Moved: The 'Add Mod' Code over to FileOriginsPage
