+++ src/NexusMods.App.UI/Pages/LoadoutGrid/LoadoutGridView.axaml
@@ -21,29 +21,16 @@
     <Grid RowDefinitions="Auto, *">
         <Border Grid.Row="0" Classes="Toolbar">
             <StackPanel>
-                <Button x:Name="DeleteModsButton">
-                    <StackPanel>
-                        <icons:UnifiedIcon Classes="Close" />
-                        <TextBlock Text="{x:Static resources:Language.LoadoutGridView__Remove}"/>
-                    </StackPanel>
-                </Button>
                 <navigation:NavigationControl x:Name="ViewModFilesButton">
                     <StackPanel>
                         <icons:UnifiedIcon Classes="FolderOutline" />
                         <TextBlock Text="{x:Static resources:Language.LoadoutGridView__View_Files}"/>
                     </StackPanel>
                 </navigation:NavigationControl>
-                <Line/>
-                <Button x:Name="AddModButton">
+                <Button x:Name="DeleteModsButton">
                     <StackPanel>
-                        <icons:UnifiedIcon Classes="PlusCircleOutline" />
-                        <TextBlock Text="{x:Static resources:Language.FileOriginsPage__Add_Mod}"/>
-                    </StackPanel>
-                </Button>
-                <Button x:Name="AddModAdvancedButton">
-                    <StackPanel>
-                        <icons:UnifiedIcon Classes="PlusCircleOutline" />
-                        <TextBlock Text="{x:Static resources:Language.FileOriginsPage__Add_Mod_Advanced}"/>
+                        <icons:UnifiedIcon Classes="Close" />
+                        <TextBlock Text="{x:Static resources:Language.LoadoutGridView__Remove}"/>
                     </StackPanel>
                 </Button>
             </StackPanel>
