+++ src/NexusMods.App.UI/Pages/LoadoutGrid/LoadoutGridViewModel.cs
@@ -8,8 +8,6 @@ using JetBrains.Annotations;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using NexusMods.Abstractions.FileStore;
-using NexusMods.Abstractions.FileStore.ArchiveMetadata;
-using NexusMods.Abstractions.Games.DTO;
 using NexusMods.Abstractions.Installers;
 using NexusMods.Abstractions.Loadouts;
 using NexusMods.Abstractions.Loadouts.Ids;
@@ -33,7 +31,6 @@ using NexusMods.App.UI.WorkspaceSystem;
 using NexusMods.Extensions.DynamicData;
 using NexusMods.Icons;
 using NexusMods.MnemonicDB.Abstractions;
-using NexusMods.Networking.Downloaders.Tasks.State;
 using NexusMods.Paths;
 using ReactiveUI;
 using ReactiveUI.Fody.Helpers;
@@ -44,12 +41,7 @@ namespace NexusMods.App.UI.Pages.LoadoutGrid;
 [UsedImplicitly]
 public class LoadoutGridViewModel : APageViewModel<ILoadoutGridViewModel>, ILoadoutGridViewModel
 {
-    private readonly IFileSystem _fileSystem;
-    private readonly ILogger<LoadoutGridViewModel> _logger;
     private readonly IConnection _conn;
-    private readonly IArchiveInstaller _archiveInstaller;
-    private readonly IFileOriginRegistry _fileOriginRegistry;
-    private readonly IServiceProvider _provider;
 
     private ReadOnlyObservableCollection<ModId> _mods;
     public ReadOnlyObservableCollection<ModId> Mods => _mods;
@@ -63,7 +55,6 @@ public class LoadoutGridViewModel : APageViewModel<ILoadoutGridViewModel>, ILoad
 
     [Reactive] public LoadoutId LoadoutId { get; set; }
     [Reactive] public string LoadoutName { get; set; } = "";
-    public GameDomain _gameDomain = GameDomain.DefaultValue;
 
     [Reactive] public ModId[] SelectedItems { get; set; } = [];
     public ReactiveCommand<NavigationInformation, Unit> ViewModContentsCommand { get; }
@@ -83,13 +74,8 @@ public class LoadoutGridViewModel : APageViewModel<ILoadoutGridViewModel>, ILoad
         IWindowManager windowManager,
         ISettingsManager settingsManager) : base(windowManager)
     {
-        _logger = logger;
-        _fileSystem = fileSystem;
         _conn = conn;
-        _archiveInstaller = archiveInstaller;
-        _fileOriginRegistry = fileOriginRegistry;
-        _provider = provider;
-        
+
         MarkdownRendererViewModel = provider.GetRequiredService<IMarkdownRendererViewModel>();
 
         _columns = new SourceCache<IDataGridColumnFactory<LoadoutColumn>, LoadoutColumn>(_ => throw new NotSupportedException());
@@ -152,15 +138,7 @@ public class LoadoutGridViewModel : APageViewModel<ILoadoutGridViewModel>, ILoad
                 .SelectMany(tuple => loadoutRepository.Revisions(tuple.First.Value))
                 .Select(loadout =>
                 {
-                    try
-                    {
-                        _gameDomain = loadout.Installation.Game.Domain;
-                    }
-                    catch (Exception)
-                    {
-                        _gameDomain = GameDomain.DefaultValue;
-                    }
-
+                    
                     var settings = settingsManager.Get<LoadoutGridSettings>();
                     var showGameFiles = settings.ShowGameFiles;
                     var showOverride = settings.ShowOverride;
@@ -191,37 +169,6 @@ public class LoadoutGridViewModel : APageViewModel<ILoadoutGridViewModel>, ILoad
                 .DisposeWith(d);
             
         });
-    }
-
-    public Task AddMod(string path) => AddMod(path, installer: null);
-
-    public Task AddModAdvanced(string path)
-    {
-        var installer = _provider.GetKeyedService<IModInstaller>("AdvancedInstaller");
-        return AddMod(path, installer);
-    }
-
-    private Task AddMod(string path, IModInstaller? installer)
-    {
-        var file = _fileSystem.FromUnsanitizedFullPath(path);
-        if (!_fileSystem.FileExists(file))
-        {
-            _logger.LogError("File {File} does not exist, not installing mod", file);
-            return Task.CompletedTask;
-        }
-
-        var _ = Task.Run(async () =>
-        {
-            var downloadId = await _fileOriginRegistry.RegisterDownload(file,
-                (tx, id) =>
-                {
-                    tx.Add(id, DownloaderState.GameDomain, _gameDomain);
-                    tx.Add(id, FilePathMetadata.OriginalName, file.FileName);
-                });
-            await _archiveInstaller.AddMods(LoadoutId, downloadId, file.FileName, token: CancellationToken.None, installer: installer);
-        });
-
-        return Task.CompletedTask;
     }
 
     public async Task DeleteMods(IEnumerable<ModId> modsToDelete, string commitMessage)
